requirements.txt
dj-database-url==0.3.0
Django==1.8
django-filter==0.9.2
djangorestframework==3.1.1
drf-nested-routers==0.9.0
gunicorn==19.3.0
Markdown==2.6.2
whitenoise==1.0.6
psycopg2==2.5.4
